requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
google-genai>=0.2.0
GitPython>=3.1.31
pandas>=2.0.0
chardet>=5.0.0
sentence-transformers>=2.2.0
boto3>=1.28.0
python-dotenv>=1.0.0

